+++ enterprise/cmd/frontend/internal/rbac/resolvers/resolver.go
@@ -9,6 +9,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/auth"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/types"
+	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
 // Resolver is the GraphQL resolver of all things related to batch changes.
@@ -32,7 +33,7 @@ func (r *Resolver) SetPermissions(ctx context.Context, args gql.SetPermissionsAr
 		return nil, err
 	}
 
-	opts := database.SyncPermissionsToRoleOpts{
+	opts := database.SetPermissionsForRoleOpts{
 		RoleID: roleID,
 	}
 
@@ -44,7 +45,7 @@ func (r *Resolver) SetPermissions(ctx context.Context, args gql.SetPermissionsAr
 		opts.Permissions = append(opts.Permissions, pID)
 	}
 
-	if err = r.db.RolePermissions().SyncPermissionsToRole(ctx, opts); err != nil {
+	if err = r.db.RolePermissions().SetPermissionsForRole(ctx, opts); err != nil {
 		return nil, err
 	}
 
@@ -111,4 +112,42 @@ func (r *Resolver) CreateRole(ctx context.Context, args *gql.CreateRoleArgs) (gq
 	}
 
 	return gql.NewRoleResolver(r.db, role), nil
+}
+
+func (r *Resolver) SetRoles(ctx context.Context, args *gql.SetRolesArgs) (*gql.EmptyResponse, error) {
+	// 🚨 SECURITY: Only site administrators can assign roles to a user.
+	// We need to get the current user any
+	if err := auth.CheckCurrentUserIsSiteAdmin(ctx, r.db); err != nil {
+		return nil, err
+	}
+
+	userID, err := gql.UnmarshalUserID(args.User)
+	if err != nil {
+		return nil, err
+	}
+
+	user, err := auth.CurrentUser(ctx, r.db)
+	if err != nil {
+		return nil, err
+	}
+
+	if user.ID == userID {
+		return nil, errors.New("cannot assign role to self")
+	}
+
+	opts := database.SetRolesForUserOpts{UserID: userID}
+
+	for _, r := range args.Roles {
+		rID, err := gql.UnmarshalPermissionID(r)
+		if err != nil {
+			return nil, err
+		}
+		opts.Roles = append(opts.Roles, rID)
+	}
+
+	if err = r.db.UserRoles().SetRolesForUser(ctx, opts); err != nil {
+		return nil, err
+	}
+
+	return &gql.EmptyResponse{}, nil
 }
